Extract IGrainTypeResolver to another class
