+++ src/Orleans.Core/Runtime/GrainInterfaceMap.cs
@@ -11,7 +11,7 @@ namespace Orleans.Runtime
     /// Internal data structure that holds a grain interfaces to grain classes map.
     /// </summary>
     [Serializable]
-    internal class GrainInterfaceMap : IGrainTypeResolver
+    internal class GrainInterfaceMap
     {
         private readonly Dictionary<string, GrainInterfaceData> typeToInterfaceData;
         private readonly Dictionary<int, GrainInterfaceData> table;
@@ -223,43 +223,7 @@ namespace Orleans.Runtime
             }
         }
 
-        public bool TryGetGrainClassData(Type interfaceType, out GrainClassData implementation, string grainClassNamePrefix)
-        {
-            implementation = null;
-            GrainInterfaceData interfaceData;
-            var typeInfo = interfaceType.GetTypeInfo();
-
-            // First, try to find a non-generic grain implementation:
-            if (this.typeToInterfaceData.TryGetValue(GetTypeKey(interfaceType, false), out interfaceData) &&
-                TryGetGrainClassData(interfaceData, out implementation, grainClassNamePrefix))
-            {
-                return true;
-            }
-
-            // If a concrete implementation was not found and the interface is generic, 
-            // try to find a generic grain implementation:
-            if (typeInfo.IsGenericType && 
-                this.typeToInterfaceData.TryGetValue(GetTypeKey(interfaceType, true), out interfaceData) &&
-                TryGetGrainClassData(interfaceData, out implementation, grainClassNamePrefix))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        public bool TryGetGrainClassData(int grainInterfaceId, out GrainClassData implementation, string grainClassNamePrefix = null)
-        {
-            implementation = null;
-            GrainInterfaceData interfaceData;
-            if (!table.TryGetValue(grainInterfaceId, out interfaceData))
-            {
-                return false;
-            }
-            return TryGetGrainClassData(interfaceData, out implementation, grainClassNamePrefix);
-        }
-
-        private string GetTypeKey(Type interfaceType, bool isGenericGrainClass)
+        internal static string GetTypeKey(Type interfaceType, bool isGenericGrainClass)
         {
             var typeInfo = interfaceType.GetTypeInfo();
             if (isGenericGrainClass && typeInfo.IsGenericType)
@@ -276,85 +240,6 @@ namespace Orleans.Runtime
             }
         }
 
-        private static bool TryGetGrainClassData(GrainInterfaceData interfaceData, out GrainClassData implementation, string grainClassNamePrefix)
-        {
-            implementation = null;
-            var implementations = interfaceData.Implementations;
-
-            if (implementations.Length == 0)
-                return false;
-
-            if (String.IsNullOrEmpty(grainClassNamePrefix))
-            {
-                if (implementations.Length == 1)
-                {
-                    implementation = implementations[0];
-                    return true;
-                }
-
-                if (interfaceData.PrimaryImplementation != null)
-                {
-                    implementation = interfaceData.PrimaryImplementation;
-                    return true;
-                }
-
-                throw new OrleansException(String.Format("Cannot resolve grain interface ID={0} to a grain class because of multiple implementations of it: {1}",
-                    interfaceData.InterfaceId, Utils.EnumerableToString(implementations, d => d.GrainClass, ",", false)));
-            }
-
-            if (implementations.Length == 1)
-            {
-                if (implementations[0].GrainClass.StartsWith(grainClassNamePrefix, StringComparison.Ordinal))
-                {
-                    implementation = implementations[0];
-                    return true;
-                }
-
-                return false;
-            }
-
-            var matches = implementations.Where(impl => impl.GrainClass.Equals(grainClassNamePrefix)).ToArray(); //exact match?
-            if (matches.Length == 0)
-                matches = implementations.Where(
-                    impl => impl.GrainClass.StartsWith(grainClassNamePrefix, StringComparison.Ordinal)).ToArray(); //prefix matches
-
-            if (matches.Length == 0)
-                return false;
-
-            if (matches.Length == 1)
-            {
-                implementation = matches[0];
-                return true;
-            }
-
-            throw new OrleansException(String.Format("Cannot resolve grain interface ID={0}, grainClassNamePrefix={1} to a grain class because of multiple implementations of it: {2}",
-                interfaceData.InterfaceId,
-                grainClassNamePrefix,
-                Utils.EnumerableToString(matches, d => d.GrainClass, ",", false)));
-        }
-
-        public bool TryGetGrainClassData(string grainImplementationClassName, out GrainClassData implementation)
-        {
-            implementation = null;
-            // have to iterate since _primaryImplementations is not serialized.
-            foreach (var interfaceData in table.Values)
-            {
-                foreach(var implClass in interfaceData.Implementations)
-                    if (implClass.GrainClass.Equals(grainImplementationClassName))
-                    {
-                        implementation = implClass;
-                        return true;
-                    }
-            }
-            return false;
-        }
-
-
-        public string GetLoadedGrainAssemblies()
-        {
-            return loadedGrainAsemblies != null ? loadedGrainAsemblies.ToStrings() : String.Empty;
-        }
-
         public void AddToUnorderedList(Type grainClass)
         {
             var grainClassTypeCode = GrainInterfaceUtils.GetGrainClassTypeCode(grainClass);
@@ -365,6 +250,16 @@ namespace Orleans.Runtime
         public bool IsUnordered(int grainTypeCode)
         {
             return unordered.Contains(grainTypeCode);
+        }
+
+        public IGrainTypeResolver GetGrainTypeResolver()
+        {
+            return new GrainTypeResolver(
+                this.typeToInterfaceData,
+                this.table,
+                this.loadedGrainAsemblies,
+                this.unordered
+                );
         }
     }
 }
